[FIX] refactor query layout
